Added internal data for calculating FGRS
Calculating the FGRS for new data requires a number of parameters
established on the full study population in othe original publication
(Kendler et al. 2021). These parameters have been extracted from
the publication appendix and the SAS code shared by the authors, for
outcome schizophrenia (SZ). The current iteration adds the data and
the code for adding these parameters to the package.