Added proper age weights to FRGS calculation
Removed a hard-coded short cut used because of a. misunderstanding the
original SAS code, b. a problem with the SZ age-cumulative incidences.
I get now that the SAS code actually implements the algorithm as described
in the paper, and have added that piece of coding to the FGRS function.

Note that the SZ cumulative incidences are still not monotonously increasing.