Added R6 class structure for FGRS data
The data structures for importing and holding the population marginal
distributions and parameters requuired for calculating the FGRS have
been improvised for the original SZ case and sloppily extended for
the BD case, with no formal structure. This commit adds a draft R6
class for collecting all required data for a phenotype into one
object, with a starter set of access functions that provide crucial
elements from which the FGRS can be calculated directly (age weights,
liabilities etc.).

Choice of R6 mostly motivted by the fact that methods are going to
be very specific, and there is no need for new generics or similar.