Re-factoring of main function FGRS for R6 data structures
Complete re-write of the FGRS main function of the package: not only has the
background data pushed into formal R6-class objects, the actual calculations
have been pushed as far as reasonable into the corresponding methods of the
FGRS_data-class.

On top of the re-write, extended debugging of the R6 methods and the example
data definitions. Especially nasty bug when using merge() to extract mean
liabilities: even with sort=FALSE, the ordering of rows in the result is
NOT guaranteed, but rather undefined, which essentially randomized
mean liabilities across relatives. This was NOT the case for the old
version of the FGRS-package, which used the same merge()-mechanism:
change between R versions? Platform-specific implementation? Anyway,
solved by using dplyr::left_join, which promises at least to *try*
to preserve the original order of the first data frame argument.