MAJOR BUG FIX: added shared genetics as factor to FGRS calculation
When calculating the total liability per relative in function FGRS, corresponding
to Step 4 in the FGRS algorithm as outlined in Kendler et al. (2021), as a
product of four terms (base liability, age correction, cohabitation, shared
genetics), for some unclear reason, the shared genetic proportion was NOT
included in the product. This has been fixed.